Make Server functionality explicit in Interpreter
- Moved all http server functionality to Interpreter
- Removed duplicate signal trapping code in src/ore.rb
- Simplified Interpreter usage in bin/ore, no longer calling duplicate signal trapping func
- User_Server is now Ore::Server<Instance as it should have been
- Switched from #decode_www_form to CGI gem to be able to parse utf-8 in request bodies
- Removed usages of that cool ▓▒░
- All tests pass

This is a part of the work of bringing features into Interpreter from some useless abstractions. Having one monolith is much simpler mental overhead
