+++ src/runtime/interpreter.rb
@@ -1,12 +1,18 @@
+require 'webrick'
+require 'cgi'
+require 'json'
+
 module Ore
 	class Interpreter
-		attr_accessor :input, :lexer, :parser, :load_standard_library, :stack, :routes, :servers, :onclick_handlers, :input_elements, :loaded_files, :source_files
+		attr_accessor :input, :lexer, :parser
+
+		attr_accessor :load_standard_library, :stack, :routes, :servers, :onclick_handlers, :input_elements, :loaded_files, :source_files, :last_output
 
 		def initialize
 			@load_standard_library = true
 			@input                 = [] # [Ore::Expression]
 			@stack                 = [] # [Ore::Scope]
-			@servers               = [] # [Ore::User_Server]
+			@servers               = [] # [Ore::Server]
 			@routes                = {} # {route: Ore::Route}
 			@loaded_files          = {} # {filename: [Ore::Expression]}
 			@source_files          = {} # {filepath: String} for error reporting
@@ -27,12 +33,42 @@ module Ore
 			@lexer.input  = source_code
 			@parser.input = @lexer.output # Tokens
 			@input        = @parser.output # Expressions
-			output
+
+			@last_output = output
+			if servers.any?
+				loop_servers
+			else
+				@last_output
+			end
 		end
 
 		def output
 			input.each.inject(nil) do |_, expr|
 				interpret expr
+			end
+		end
+
+		def loop_servers
+			begin
+				keep_running = true
+
+				trap_fn = Proc.new do
+					puts @last_output
+					puts Ore::Ascii.dim "Shutting down..."
+					keep_running = false
+					Thread.main.raise Interrupt
+				end
+				Signal.trap 'INT', trap_fn
+				Signal.trap 'TERM', trap_fn
+
+				while keep_running
+					@servers.each do |server|
+						puts "Ore Server `#{server.server_instance.name}` started at http://localhost:#{server.port}"
+						server.server_thread&.join
+					end
+				end
+			ensure
+				@servers.each { |s| stop_server s }
 			end
 		end
 
@@ -1095,6 +1131,8 @@ module Ore
 			interp_string expr.body
 		end
 
+		# @param expr [Ore::Route_Expr]
+		# @return Ore::Route
 		def interp_route expr
 			func = interpret expr.expression
 
@@ -1434,12 +1472,14 @@ module Ore
 					raise Ore::Invalid_Start_Directive_Argument.new(expr, self)
 				end
 
-				collected_routes = collect_routes_from_instance server_instance
-				user_server      = Ore::User_Server.new server_instance, self, collected_routes
-				servers << user_server
+				server                 = Ore::Server.new
+				server.server_instance = server_instance
+				server.port            = Integer(server_instance.get(:port) || Ore::Server::DEFAULT_PORT)
+				server.routes          = collect_routes_from_instance server_instance
+				servers << server
 
-				user_server.start
-				user_server
+				start_server server
+				server
 			when 'connect'
 				database = interpret expr.expression
 				database.create_connection!
@@ -1575,6 +1615,239 @@ module Ore
 			else
 				raise Ore::Interpret_Expr_Not_Implemented.new(expr, self)
 			end
+		end
+
+		def start_server server
+			webrick = WEBrick::HTTPServer.new Port:      server.port,
+			                                  Logger:    WEBrick::Log.new("/dev/null"),
+			                                  AccessLog: []
+
+			webrick.mount_proc '/onclick/' do |req, res|
+				puts Ore::Ascii.dim "#{'DOM'.rjust(7, ' ')} #{req.path}"
+				handle_request server, req, res
+			end
+
+			webrick.mount_proc '' do |req, res|
+				handle_request server, req, res
+			end
+
+			server.webrick_server = webrick
+			server.server_thread  = Thread.new { webrick.start }
+			server.server_thread
+		end
+
+		def stop_server server
+			server.webrick_server&.shutdown
+			Thread.kill server.server_thread if server.server_thread
+		end
+
+		def handle_request server, request, response
+			path_string  = request.path
+			query_string = request.query_string
+			http_method  = request.request_method.downcase
+			path_parts   = request.path.split('/').reject { _1.empty? }
+			body_hash    = CGI.parse(request.body || "").transform_values(&:first)
+			headers_hash = request.header.to_h
+
+			req_info = Ascii.dim ""
+			unless body_hash.empty?
+				req_info = req_info.prepend Ascii.green
+			end
+			req_info << Ascii.dim(http_method.upcase.rjust(7, " "))
+			req_info << " "
+			req_info << Ascii.reset(path_string.gsub("/", "#{Ascii.dim('/')}#{Ascii.reset}"))
+			unless body_hash.empty?
+				req_info << " #{Ascii.dim body_hash}"
+			end
+			puts req_info
+
+			if path_string.start_with?("/onclick/")
+				object_id = path_parts.last.to_i
+				handler   = onclick_handlers[object_id]
+				if handler
+					begin
+						if request.body && !request.body.empty?
+							json_body = JSON.parse request.body rescue {}
+							inputs = json_body['inputs'] || {}
+							inputs.each do |element_id, value|
+								input_instance = input_elements[element_id.to_i]
+								input_instance.declare 'value', value if input_instance
+							end
+						end
+
+						route             = Ore::Route.new
+						route.handler     = handler
+						route.param_names = []
+
+						req = build_ore_request path_string, http_method, body_hash, parse_query_string(query_string), {}, headers_hash
+						res = build_ore_response response
+
+						interp_route_handler route, req, res
+
+						component = handler.enclosing_scope
+						if component.is_a?(Ore::Instance) && component.declarations['render']
+							new_html = render_dom_to_html component
+							html_id  = component.declarations['html_id']
+
+							response.status             = 200
+							response['Content-Type']    = 'text/html'
+							response['X-Ore-Target-Id'] = html_id if html_id
+							response.body               = new_html
+							return
+						end
+					rescue => e
+						warn "\n[Ore Onclick Error] #{e.class}: #{e.message}"
+						warn e.backtrace.first(10).map { |line| "  #{line}" }.join("\n")
+						warn ""
+
+						plain_message   = e.message.gsub(/\e\[\d+(?:;\d+)*m/, '')
+						response.status = 500
+						response.body   = "Internal Server Error\n#{plain_message}"
+						return
+					end
+				end
+			end
+
+			if cookie = request.cookies.find { _1.name == BROWSER_VIEW_SIZE }
+				parts = cookie.value.split 'x'
+				size  = { width: parts[0].to_i, height: parts[1].to_i }
+				stack.last.declare BROWSER_VIEW_SIZE, size
+			end
+
+			route_function = match_route http_method, path_parts, server.routes
+
+			if route_function
+				url_params   = extract_url_params path_parts, route_function
+				query_params = parse_query_string query_string
+
+				req = build_ore_request path_string, http_method, body_hash, query_params, url_params, headers_hash
+
+				begin
+					res    = build_ore_response response
+					result = interp_route_handler route_function, req, res, url_params, server_instance: server.server_instance
+
+					response.status = res.declarations['status']
+					response.body   = res.declarations['body']
+					res.declarations['headers'].each { |k, v| response.header[k] = v }
+
+					if response.body.to_s =~ /<html|<body|<head/i
+						response.body.prepend "<!DOCTYPE html>"
+
+						dom_js     = File.read 'src/runtime/dom.js'
+						script_tag = "<script>#{dom_js}</script>"
+
+						body_str = response.body.to_s
+						if body_str.include?('<head>')
+							response.body = body_str.sub('<head>', '<head>' + script_tag)
+						elsif body_str.include?('<body>')
+							response.body = body_str.sub('<body>', '<body>' + script_tag)
+						else
+							response.body = script_tag + body_str
+						end
+					end
+
+					result
+
+				rescue WEBrick::HTTPStatus::Status => e
+					raise e
+
+				rescue => e
+					warn "\n[Ore Server Error] #{e.class}: #{e.message}"
+					warn e.backtrace.first(10).map { |line| "  #{line}" }.join("\n")
+					warn ""
+
+					plain_message   = e.message.gsub(/\e\[\d+(?:;\d+)*m/, '')
+					plain_backtrace = e.backtrace.map { |line| line.gsub(/\e\[\d+(?:;\d+)*m/, '') }
+
+					response.status = 500
+					response.body   = <<~HTML
+					    <h1>500 Internal Server Error</h1>
+					    <h2>#{e.class}</h2>
+					    <pre>#{plain_message}</pre>
+					    <h3>Backtrace</h3>
+					    <pre>#{plain_backtrace.join("\n")}</pre>
+					HTML
+					response.header['Content-Type'] = 'text/html; charset=utf-8'
+				end
+			else
+				response.status = 404
+				response.body   = <<~HTML
+				    <h1>404 Not Found</h1>
+				    <p>No route matches #{http_method.upcase} #{path_string}</p>
+				    <hr>
+				    <h3>Available Routes:</h3>
+				    <ul>
+				    	#{server.routes.values.map { |r| "<li>#{r.http_method.value.upcase} /#{r.path}</li>" }.join("\n")}
+				    </ul>
+				HTML
+				response.header['Content-Type'] = 'text/html; charset=utf-8'
+			end
+		end
+
+		def match_route http_method, path_parts, routes
+			routes.values.find do |route|
+				next unless route.http_method.value == http_method
+				next unless route.parts.count == path_parts.count
+
+				path_parts.zip(route.parts).all? do |req_part, route_part|
+					(req_part == route_part) || (route_part.start_with?(':'))
+				end
+			end
+		end
+
+		def extract_url_params path_parts, route
+			url_params = {}
+			path_parts.zip(route.parts).each do |req_part, route_part|
+				if route_part.start_with? ':'
+					param_name                    = route_part[1..-1]
+					url_params[param_name]        = req_part
+					url_params[param_name.to_sym] = req_part
+				end
+			end
+			url_params
+		end
+
+		def parse_query_string query_string
+			query_params = {}
+			if query_string
+				query_string.split('&').each do |pair|
+					key, value               = pair.split '=', 2
+					query_params[key]        = CGI.unescape(value || '')
+					query_params[key.to_sym] = CGI.unescape(value || '')
+				end
+			end
+			query_params
+		end
+
+		def build_ore_request path_string, http_method, body_hash, query_params, url_params, headers_hash
+			req          = Ore::Request.new
+			body_dict    = Ore::Dictionary.new body_hash
+			query_dict   = Ore::Dictionary.new query_params
+			params_dict  = Ore::Dictionary.new url_params
+			headers_dict = Ore::Dictionary.new headers_hash
+			link_instance_to_type req, 'Request'
+			link_instance_to_type body_dict, 'Dictionary'
+			link_instance_to_type query_dict, 'Dictionary'
+			link_instance_to_type params_dict, 'Dictionary'
+			link_instance_to_type headers_dict, 'Dictionary'
+			req.declarations['path']              = path_string
+			req.declarations['method']            = http_method
+			req.declarations['query']             = query_dict
+			req.declarations['params']            = params_dict
+			req.declarations['headers']           = headers_dict
+			req.declarations['body']              = body_dict
+			req.declarations['body'].declarations = body_hash
+			req
+		end
+
+		def build_ore_response webrick_response
+			res                                  = Ore::Response.new webrick_response
+			res.declarations['webrick_response'] = webrick_response
+			res.declarations['status']           = 200
+			res.declarations['headers']          = {}
+			res.declarations['body']             = ''
+			link_instance_to_type res, 'Response'
+			res
 		end
 	end
 end
